Define `Stream` interface
- rewrite the whole `Stream` interface from Java to Go,
including the methods coming from `BaseStream`,
the whole interface is in `stream/iface.go`
- define the static methods from the `Stream` interface in Java
as package-wide methods of the `stream` package - `stream/static.go`.
No implementation for these methods yes.
- define the methods that in Java are part of the `Stream` interface,
but can't be in Go as package methods in `stream/extracted.go`.
No implementation for these methods yet.
- define the other types that are used by the Java interface
in `stream/types.go`.
Some of these types are just placeholders in order for the code to compile
and they will be implemented later.